go/consensus/cometbft: Move methods EventsFromCometBFT
