+++ go/consensus/cometbft/registry/registry.go
@@ -3,7 +3,6 @@ package registry
 
 import (
 	"context"
-	"errors"
 	"fmt"
 
 	cmtabcitypes "github.com/cometbft/cometbft/abci/types"
@@ -11,13 +10,11 @@ import (
 	cmttypes "github.com/cometbft/cometbft/types"
 	"github.com/eapache/channels"
 
-	"github.com/oasisprotocol/oasis-core/go/common/crypto/hash"
 	"github.com/oasisprotocol/oasis-core/go/common/entity"
 	"github.com/oasisprotocol/oasis-core/go/common/logging"
 	"github.com/oasisprotocol/oasis-core/go/common/node"
 	"github.com/oasisprotocol/oasis-core/go/common/pubsub"
 	consensus "github.com/oasisprotocol/oasis-core/go/consensus/api"
-	eventsAPI "github.com/oasisprotocol/oasis-core/go/consensus/api/events"
 	tmapi "github.com/oasisprotocol/oasis-core/go/consensus/cometbft/api"
 	app "github.com/oasisprotocol/oasis-core/go/consensus/cometbft/apps/registry"
 	"github.com/oasisprotocol/oasis-core/go/registry/api"
@@ -287,87 +284,6 @@ func (sc *ServiceClient) DeliverEvent(ctx context.Context, height int64, tx cmtt
 	}
 
 	return nil
-}
-
-// EventsFromCometBFT extracts registry events from CometBFT events.
-func EventsFromCometBFT(
-	tx cmttypes.Tx,
-	height int64,
-	tmEvents []cmtabcitypes.Event,
-) ([]*api.Event, []*NodeListEpochInternalEvent, error) {
-	var txHash hash.Hash
-	switch tx {
-	case nil:
-		txHash.Empty()
-	default:
-		txHash = hash.NewFromBytes(tx)
-	}
-
-	var events []*api.Event
-	var nodeListEvents []*NodeListEpochInternalEvent
-	var errs error
-	for _, tmEv := range tmEvents {
-		// Ignore events that don't relate to the registry app.
-		if tmEv.GetType() != app.EventType {
-			continue
-		}
-
-		for _, pair := range tmEv.GetAttributes() {
-			key := pair.GetKey()
-			val := pair.GetValue()
-
-			switch {
-			case eventsAPI.IsAttributeKind(key, &api.NodeListEpochEvent{}):
-				// Node list epoch event (value is ignored).
-				nodeListEvents = append(nodeListEvents, &NodeListEpochInternalEvent{Height: height})
-			case eventsAPI.IsAttributeKind(key, &api.RuntimeStartedEvent{}):
-				// Runtime started event.
-				var e api.RuntimeStartedEvent
-				if err := eventsAPI.DecodeValue(val, &e); err != nil {
-					errs = errors.Join(errs, fmt.Errorf("registry: corrupt RuntimeStarted event: %w", err))
-					continue
-				}
-
-				events = append(events, &api.Event{Height: height, TxHash: txHash, RuntimeStartedEvent: &e})
-			case eventsAPI.IsAttributeKind(key, &api.RuntimeSuspendedEvent{}):
-				// Runtime suspended event.
-				var e api.RuntimeSuspendedEvent
-				if err := eventsAPI.DecodeValue(val, &e); err != nil {
-					errs = errors.Join(errs, fmt.Errorf("registry: corrupt RuntimeSuspended event: %w", err))
-					continue
-				}
-
-				events = append(events, &api.Event{Height: height, TxHash: txHash, RuntimeSuspendedEvent: &e})
-			case eventsAPI.IsAttributeKind(key, &api.EntityEvent{}):
-				// Entity event.
-				var e api.EntityEvent
-				if err := eventsAPI.DecodeValue(val, &e); err != nil {
-					errs = errors.Join(errs, fmt.Errorf("registry: corrupt Entity event: %w", err))
-					continue
-				}
-
-				events = append(events, &api.Event{Height: height, TxHash: txHash, EntityEvent: &e})
-			case eventsAPI.IsAttributeKind(key, &api.NodeEvent{}):
-				// Node event.
-				var e api.NodeEvent
-				if err := eventsAPI.DecodeValue(val, &e); err != nil {
-					errs = errors.Join(errs, fmt.Errorf("registry: corrupt Node event: %w", err))
-					continue
-				}
-
-				events = append(events, &api.Event{Height: height, TxHash: txHash, NodeEvent: &e})
-			case eventsAPI.IsAttributeKind(key, &api.NodeUnfrozenEvent{}):
-				// Node unfrozen event.
-				var e api.NodeUnfrozenEvent
-				if err := eventsAPI.DecodeValue(val, &e); err != nil {
-					errs = errors.Join(errs, fmt.Errorf("registry: corrupt NodeUnfrozen event: %w", err))
-					continue
-				}
-				events = append(events, &api.Event{Height: height, TxHash: txHash, NodeUnfrozenEvent: &e})
-			}
-		}
-	}
-	return events, nodeListEvents, errs
 }
 
 func (sc *ServiceClient) getNodeList(ctx context.Context, height int64) (*api.NodeList, error) {
